+++ click_extra/tests/test_commands.py
@@ -20,7 +20,6 @@ options, and how they interact with each others."""
 from __future__ import annotations
 
 import re
-import sys
 from textwrap import dedent
 
 import click
@@ -207,11 +206,10 @@ def test_integrated_version_value(invoke, all_command_cli):
     result = invoke(all_command_cli, "--version", color=False)
     assert result.exit_code == 0
 
-    regex_output = r"command-cli1, version 2021.10.08\n"
-    # XXX Temporarily skip displaying environment details for Python >= 3.10 while we
-    # wait for https://github.com/mahmoud/boltons/issues/294 to be released upstream.
-    if sys.version_info[:2] < (3, 10):
-        regex_output += r"{'.+'}\n"
+    regex_output = (
+        r"command-cli1, version 2021.10.08\n"
+        r"{'.+'}\n"
+    )
     assert re.fullmatch(regex_output, result.output)
 
     assert not result.stderr