+++ click_extra/tests/test_version.py
@@ -17,7 +17,6 @@
 from __future__ import annotations
 
 import re
-import sys
 
 import click
 import pytest
@@ -42,12 +41,11 @@ def test_standalone_version_option_with_env_info(invoke):
     result = invoke(color_cli2, "--version", color=True)
     assert result.exit_code == 0
 
-    regex_output = r"\x1b\[97mcolor-cli2\x1b\[0m, version \x1b\[32m1.2.3.4"
-    # XXX Temporarily skip displaying environment details for Python >= 3.10 while we
-    # wait for https://github.com/mahmoud/boltons/issues/294 to be released upstream.
-    if sys.version_info[:2] < (3, 10):
-        regex_output += r"\x1b\[0m\x1b\[90m\n{'.+'}"
-    regex_output += r"\x1b\[0m\n"
+    regex_output = (
+        r"\x1b\[97mcolor-cli2\x1b\[0m, version \x1b\[32m1.2.3.4"
+        r"\x1b\[0m\x1b\[90m\n{'.+'}"
+        r"\x1b\[0m\n"
+    )
     assert re.fullmatch(regex_output, result.output)
 
     assert not result.stderr
@@ -85,12 +83,11 @@ def test_integrated_version_option_precedence(invoke, params):
     result = invoke(color_cli4, "--version", params, color=True)
     assert result.exit_code == 0
 
-    regex_output = r"\x1b\[97mcolor-cli4\x1b\[0m, version \x1b\[32m1.2.3.4"
-    # XXX Temporarily skip displaying environment details for Python >= 3.10 while we
-    # wait for https://github.com/mahmoud/boltons/issues/294 to be released upstream.
-    if sys.version_info[:2] < (3, 10):
-        regex_output += r"\x1b\[0m\x1b\[90m\n{'.+'}"
-    regex_output += r"\x1b\[0m\n"
+    regex_output = (
+        r"\x1b\[97mcolor-cli4\x1b\[0m, version \x1b\[32m1.2.3.4"
+        r"\x1b\[0m\x1b\[90m\n{'.+'}"
+        r"\x1b\[0m\n"
+    )
     assert re.fullmatch(regex_output, result.output)
 
     assert not result.stderr